refactor incoming code -> sampler, drop unbuffered sampling (disk writing)
