+++ python/cugraph-pyg/cugraph_pyg/sampler/io/__init__.py
@@ -12,5 +12,4 @@
 # limitations under the License.
 
 
-from .reader import BufferedSampleReader, DistSampleReader
-from .writer import DistSampleWriter
+from .reader import BufferedSampleReader
